Move cmd to internal package
