+++ cmd/coder/main.go
@@ -9,8 +9,8 @@ import (
 	"os"
 	"runtime"
 
+	"cdr.dev/coder-cli/internal/cmd"
 	"cdr.dev/coder-cli/internal/x/xterminal"
-	"github.com/spf13/cobra"
 
 	"go.coder.com/flog"
 )
@@ -35,78 +35,11 @@ func main() {
 	}
 	defer xterminal.Restore(os.Stdout.Fd(), stdoutState)
 
-	app := &cobra.Command{
-		Use:     "coder",
-		Short:   "coder provides a CLI for working with an existing Coder Enterprise installation",
-		Version: fmt.Sprintf("%s %s %s/%s", version, runtime.Version(), runtime.GOOS, runtime.GOARCH),
-	}
+	app := cmd.Make()
+	app.Version = fmt.Sprintf("%s %s %s/%s", version, runtime.Version(), runtime.GOOS, runtime.GOARCH)
 
-	app.AddCommand(
-		makeLoginCmd(),
-		makeLogoutCmd(),
-		makeShellCmd(),
-		makeUsersCmd(),
-		makeConfigSSHCmd(),
-		makeSecretsCmd(),
-		makeEnvsCommand(),
-		makeSyncCmd(),
-		makeURLCmd(),
-		completionCmd,
-	)
 	err = app.ExecuteContext(ctx)
 	if err != nil {
 		os.Exit(1)
 	}
-}
-
-// reference: https://github.com/spf13/cobra/blob/master/shell_completions.md
-var completionCmd = &cobra.Command{
-	Use:   "completion [bash|zsh|fish|powershell]",
-	Short: "Generate completion script",
-	Long: `To load completions:
-
-Bash:
-
-$ source <(yourprogram completion bash)
-
-# To load completions for each session, execute once:
-Linux:
-  $ yourprogram completion bash > /etc/bash_completion.d/yourprogram
-MacOS:
-  $ yourprogram completion bash > /usr/local/etc/bash_completion.d/yourprogram
-
-Zsh:
-
-# If shell completion is not already enabled in your environment you will need
-# to enable it.  You can execute the following once:
-
-$ echo "autoload -U compinit; compinit" >> ~/.zshrc
-
-# To load completions for each session, execute once:
-$ yourprogram completion zsh > "${fpath[1]}/_yourprogram"
-
-# You will need to start a new shell for this setup to take effect.
-
-Fish:
-
-$ yourprogram completion fish | source
-
-# To load completions for each session, execute once:
-$ yourprogram completion fish > ~/.config/fish/completions/yourprogram.fish
-`,
-	DisableFlagsInUseLine: true,
-	ValidArgs:             []string{"bash", "zsh", "fish", "powershell"},
-	Args:                  cobra.ExactValidArgs(1),
-	Run: func(cmd *cobra.Command, args []string) {
-		switch args[0] {
-		case "bash":
-			cmd.Root().GenBashCompletion(os.Stdout)
-		case "zsh":
-			cmd.Root().GenZshCompletion(os.Stdout)
-		case "fish":
-			cmd.Root().GenFishCompletion(os.Stdout, true)
-		case "powershell":
-			cmd.Root().GenPowerShellCompletion(os.Stdout)
-		}
-	},
 }
